Poursuit adaptation float

+++ lien.h
@@ -11,7 +11,7 @@ class lien
 public:
     lien(matflux & w_mflux, std::vector<commune> & w_vcom) : mflux(w_mflux), vcom(w_vcom) {}
 
-    virtual double calcul_elem(int numdc_l, int numdc_c, int nb, int nb_d) = 0;
+    virtual double calcul_elem(int numdc_l, int numdc_c, double nb, double nb_d) = 0;
     virtual double val_init() const = 0;
     virtual double val_stop() const = 0;
     virtual bool is_sym() const = 0;
